requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
requests
psycopg2-binary
